Major refactor and new pyvisa simulator
The structure of the UI class has been changed drastically:
It is now split into multiple subclasses which are responsible
for handling part of the GUI and are inherited by the main
window class.

A new pyvisa simulator has been implemented as well,
which enables development without access to relevant devices.