Refactoring of the Plotter class and more simulator features
The simulator now returns an appropriate amount of binary data
when we're trying to read the buffers of an SR830.
Pyvisa-sim was considered for replacing the homebrew simulator,
but it appears that it cannot handle the binary data transfer.

Saving and quick saving are now implemented for the plot.
Quick saving may be done to a user-defined directory in the future.