ODMRWaterfall: better PSU control
Before starting a sweep, the program will wait for
the current to settle and read it back into the array.
TODO: Make the wait time configurable

When shutting down the PSU, the current is stepped down
in small increments before the output is disabled in order
to prevent the inductor from triggering reverse voltage protection.
TODO: Make the step size and speed configurable